+++ cinShared.py
@@ -1,0 +1,16 @@
+import re
+
+urlRegex = 'http[s]?://(?:[a-zA-Z]|[0-9]|[$-_@.&+]|[!*\(\),]|(?:%[0-9a-fA-F][0-9a-fA-F]))+'
+
+
+
+def containsAny(textToCheck: str, texts):
+    # if textToCheck contains any of texts, return the length of the matching text, which is truthy unless text is ""
+    # else, return false
+    for text in texts:
+        if text in textToCheck.lower():
+            return len(text)
+    return False
+
+def getURLs(string):
+    return re.findall(urlRegex, string)